Add imports to generated functypes file
To properly add imports, the ident used in the generated types need to be
recorded in the package's TypeInfo. There is a risk of two files in the same
package trying to import different packages under the same name. The compiler
will stop and report the issue if that happens, but we should consider adding a
pre-pass to normalize all imports in the package (remove dot imports, qualify
them all with unique names, etc.)

Fixes #57